Add files via upload
A Seaborn heatmap is a graphical representation of data in a matrix format, where values in a dataset are represented as colors in a two-dimensional space. It is an effective way to visualize relationships between variables, especially when working with numerical data. Seaborn is a Python data visualization library built on top of Matplotlib that makes it easy to create aesthetically pleasing and informative heatmaps.